requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.3.0
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
aiohttp>=3.8.0
python-multipart>=0.0.6
streamlit>=1.28.0
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Advanced Features Dependencies
redis>=5.0.0
redis[hiredis]>=5.0.0  # High-performance Redis client
slowapi>=0.1.9
websockets>=11.0.0
cohere>=4.37.0
google-generativeai>=0.3.2
anthropic>=0.7.8
tiktoken>=0.5.2
plotly>=5.17.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
celery[redis]>=5.3.0
apscheduler>=3.10.4
jinja2>=3.1.2
httpx>=0.25.0
cachetools>=5.3.0

# Performance Optimization Dependencies
asyncio-throttle>=1.0.2  # Rate limiting for async operations
aiocache>=0.12.2  # Advanced async caching
aioredis>=2.0.1  # Async Redis client (alternative)
orjson>=3.9.0  # Fast JSON serialization
uvloop>=0.19.0  # High-performance event loop (Unix only)

# Data Science Dependencies (for advanced analytics)
numpy>=1.24.0
scipy>=1.11.0

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development Dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Security Dependencies
safety>=2.3.0
bandit>=1.7.0

# Performance Testing
locust>=2.17.0

# Production Dependencies
gunicorn>=21.2.0
